[AS3935MI] Fix switching display LCO frequencies via I2C
What I noticed on ESP32 is that whenever I try to send I2C calls to read a register to turn off displaying the LCO frequency on the IRQ pin, that the read is retried a number of times but the sensor reports NAK (as displayed by the protocol analysis of the logic analyser)

Thus Wire.read() returns -1. This is 0xFF in uint8_t.

So the function to then patch the bits to turn off the LCO leaves the other bits set to 1 which then enables both the RCO signals.

These then show a nice combined aliasing signal of both xRCO frequencies on the IRQ pin.

‌So what I now do is that I keep a cached copy of the antenna setting and use that to write directly to register 8.